more code cleanup

+++ kernel/cstdlib/strings/stringintconversions.c
@@ -4,7 +4,7 @@
 int64_t stringToInt64(const char* str) {
     int64_t result = 0;
     int64_t sign = 1;
-    int64_t i = 0;
+    size_t i = 0;
 
     if (str[0] == '-') {
         sign = -1;
@@ -44,7 +44,7 @@ uint64_t stringToUnsignedInt64(const char* str) {
 int32_t stringToInt32(const char* str) {
     int32_t result = 0;
     int32_t sign = 1;
-    int32_t i = 0;
+    size_t i = 0;
 
     if (str[0] == '-') {
         sign = -1;
@@ -81,7 +81,7 @@ uint32_t stringToUnsignedInt32(const char* str) {
 int16_t stringToInt16(const char* str) {
     int16_t result = 0;
     int16_t sign = 1;
-    int16_t i = 0;
+    size_t i = 0;
 
     if (str[0] == '-') {
         sign = -1;
@@ -118,7 +118,7 @@ uint16_t stringToUnsignedInt16(const char* str) {
 int8_t stringToInt8(const char* str) {
     int8_t result = 0;
     int8_t sign = 1;
-    int8_t i = 0;
+    size_t i = 0;
 
     if (str[0] == '-') {
         sign = -1;
@@ -155,7 +155,7 @@ uint8_t stringToUnsignedInt8(const char* str) {
 
 void UIntToString(unsigned int num, char* str) {
     uint8_t* NumberBuffer = (uint8_t*)&num;
-    int i;
+    size_t i;
     for (i = 0; i < sizeof(int); i++) {
         str[sizeof(int) - 1 - i] = NumberBuffer[i];
     }
